[Refactor] TopAppBarComponent 재사용 요소 추가(CartAppBarComponent -> TopAppBarComponent)
