+++ app/src/main/java/com/example/toucheeseapp/ui/screens/tab_Home/CartScreen.kt
@@ -5,6 +5,9 @@ import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.automirrored.filled.ArrowBack
+import androidx.compose.material.icons.filled.ShoppingCart
 import androidx.compose.material3.*
 import androidx.compose.runtime.*
 import androidx.compose.ui.Alignment
@@ -94,10 +97,11 @@ fun CartScreen(
         topBar = {
             TopAppBarComponent(
                 title = "장바구니",
+                leadingIcon = Icons.AutoMirrored.Default.ArrowBack,
                 showLeadingIcon = true,
                 showTrailingIcon = false,
                 onClickLeadingIcon = onBackClick,
-                onClickTrailingICon = { /* 필요시 구현 */}
+                onClickTrailingIcon = { /* 필요시 구현 */}
             )
         },
         bottomBar = {
