+++ app/src/main/java/com/example/toucheeseapp/ui/screens/tab_Home/OrderPayScreen.kt
@@ -76,7 +76,7 @@ fun OrderPayScreen(
                 showLeadingIcon = true,
                 showTrailingIcon = false,
                 onClickLeadingIcon = onBackClick,
-                onClickTrailingICon = { /* 필요시 구현 */ }
+                onClickTrailingIcon = { /* 필요시 구현 */ }
             )
         },
         bottomBar = {
